perf(celt): SIMD denormaliseBands requant via i32.GainQ31 (#49)
* perf(celt): SIMD denormaliseBands requant via simd i32.GainQ31

Replace the per-band inner requant loop in denormaliseBands with a single
fused i32.GainQ31 pass. The loop

    fb[k] = PSHR32(MULT32_32_Q31(SHL32(xv, 30-normShift), g), shift)

maps exactly to GainQ31(fb, xb, g, 30-normShift, shift): preShift is the
compile-time constant 6 (normShift = 24) and shift is clamped to [0,30]
just above the loop, so both stay inside GainQ31's [0,31] domain and it
never panics. GainQ31 is documented bit-identical to that fixedmath
composition (wrapping int32 at every stage, no relaxed tier).

Follows the existing haar1/pitch scalar-reference + SIMD-dispatch +
differential-test convention:
  - denormalise_ref.go:  denormaliseGainGeneric, the frozen scalar reference
  - denormalise_simd.go: denormaliseGain, the i32.GainQ31 form
  - denormalise_simd_test.go: a matrix / extremes / in-place / fuzz
    differential test that re-asserts scalar-vs-SIMD byte identity on every
    build, so the guard holds even in CI jobs that do not build the cgo
    refc oracle.

GainQ31 landed in simd after v1.6.0, so this bumps the pin from v1.5.0 to
v1.8.0. That bump is purely additive on the simd side (the ScaleQ31,
Butterfly, i16 pitch and cint FFT kernels this module already uses are
byte-identical between the two versions), and every existing kernel keeps
passing its differential test under the new pin.

Bit-exact: the refc differential gate passes on amd64 (AVX2) and arm64
(NEON), plus ~3M fuzz executions and the per-build differential test.

Performance (decode; arm64 measured interleaved, core-pinned, n=12, p<0.001):
  - arm64 (Cortex-A76): 10 ms and 20 ms frames 1.3 to 1.65 percent faster;
    5 ms neutral to 0.5 percent faster; 2.5 ms frames about 0.6 percent
    slower (narrowest bands, where the per-band call is not yet amortized).
    Geomean +0.66 percent throughput. The kernel itself is 1.5 to 1.66x on
    the wide bands (n >= 48), which hold most of the samples.
  - amd64: end-to-end neutral (within noise). The simd library gates its
    AVX2 path at n >= 160, so every band but the widest routes through the
    library's scalar path, which matches the old inlined loop; on amd64 the
    change is a no-cost code-path unification, and the real win is on arm64.

No length gate is added: the simd library already gates internally
(minAVX2GainQ31 = 160, minNEONGainQ31 = 4), and the arm64 net win already
absorbs the small-band cost.

* fix(celt): keep SIMD benchmark seeds 32-bit portable

int32(i*2654435761 + 1) does not compile on 32-bit platforms (GOARCH=386,
arm): the untyped constant 2654435761 overflows a 32-bit int. Multiply in
uint32 instead (matching pitch_simd_test.go's existing idiom) so the seed
wraps in uint32 and stays portable. Fixes the new denormaliseGain benchmark
and the sibling haar1 benchmark it was copied from.

Verified with GOARCH=386 and GOARCH=arm go vet.

+++ internal/celt/haar1_simd_test.go
@@ -116,7 +116,7 @@ func BenchmarkHaar1(b *testing.B) {
 		}
 		seed := make([]int32, n)
 		for i := range seed {
-			seed[i] = int32(i*2654435761 + 1)
+			seed[i] = int32(uint32(i)*2654435761 + 1)
 		}
 		b.Run(fmt.Sprintf("scalar/N%d_s%d", c.N0, c.stride), func(b *testing.B) {
 			buf := append([]int32(nil), seed...)
